AppNavigator [nfc]: Parameterize by initial route name/params.
Keeping (for now) the values that had been used before, i.e.,
'loading' for `initialRouteName` and nothing for
`initialRouteParams`.

`AppNavigator` is now computed at the top level of
ZulipAppContainer.js, rather than AppNavigator.js. In an upcoming
commit, we'll further defer the computation to within the
constructor of a new React component, `ZulipAppContainer`.

The awkwardness of doing that will disappear in the upcoming React
Navigation v5 upgrade, in which the new component-based API will do
away with `create*` functions in favor of plain React components to
be invoked in JSX.