unread tests: Fix some impossible data.
One consequence of building test data manually through the internal
representation is that sometimes it doesn't make any sense!

These message IDs collide with each other; and then in the "huddles"
(i.e. group-PMs) part of the data, the keys are all supposed to
contain the self user, so they must have a user ID in common.  If we
say user 1 is self, that's consistent with the 1:1 PMs side of this
data, so we pick that.

Discovered these while converting this data to a more structured
form where these inconsistencies can't happen.  Fix them first,
to simplify the conversion.