ZulipAppContainer [nfc]: Further defer calling `createAppContainer`.
As foreshadowed in a recent commit, the call to `createAppContainer`
is no longer in `ZulipAppContainer`'s toplevel, but rather in the
constructor of the `ZulipAppContainer` React component, and we store
it in an instance property on the component.

The difference is that now, at its current callsite, it's possible
to be aware of the things held in Redux that would allow us to set
`initialRouteName` and `initialRouteParams`, appropriately, before
any navigation config is established. This will let us remove the
awkward `InitialNavigationDispatcher` component and get us using the
`react-navigation`-appropriate levers for choosing the initial route
and params.