msglist [nfc]: Improve inbound/outbound event variable names (1/2).
The most important thing to know about any of these "events" is
whether it's being sent from the React Native context to the context
inside the WebView, or vice versa. So, make the variable names clear
about this:

MessageListEvent -> WebViewOutboundEvent

WebViewUpdateEvent -> WebViewInboundEvent
getUpdateEvents -> generateInboundEvents

Done with a find-and-replace-all in VSCode.