user [nfc]: Rename old UserItem as UserItemRaw, for its one caller.
The thing we're doing here in the autocomplete suggestions probably
isn't *great*, or the ideal UI, and certainly makes a bit of a mess
in the code.  But doing better is a different kind of change from the
refactoring we're engaged in in this series.  For now, just give an
appropriate name and comments, to keep the mess isolated to this one
callsite.

We'll next rename things so that the new UserItemById is simply
UserItem, the default export; that'll be a separate commit.