user types: Propagate UserId type through outbound typing-status.
The data we're passing to this code comes from PM recipients,
where we've already marked the user IDs as type `UserId`.  So
when we tighten the type here, Flow already accepts that.

This in turn causes Flow to see that this code's local helper
`getRecipients` operates on `UserId`s.  It passes the user IDs
to one of our Redux selectors; so in an upcoming commit, when we
tighten those selectors to expect `UserId`, Flow will see that
this change is in fact required.