AnimatedRotateComponent [nfc]: Add simple type annotation.
With the new Flow version (0.122.0) coming along with the React
Native v0.63 upgrade (#4245), this would be flagged with a "Could
not decide which case to select" error.

I'm not sure why it has trouble distinguishing this array of strings
from a possible array of numbers, but it's simple enough to just
annotate it as an array of strings, so, might as well.