messagesReducer [nfc]: Fix a sketchy null check.
This causes Flow to mark the `$FlowFixMe` on
`EventUpdateMessageAction.orig_subject?` as an "unused suppression
comment", so, remove it. In fact, I don't think a `$FlowFixMe`
really belongs here; there's no question that `orig_subject` is
sometimes missing, if only because private messages don't have
topics. The problem is in working out when it's there and when it's
not.

As I point out in the ongoing discussion [1] linked from that TODO,
there's a particular line [2] in the server code at current `master`
that I believe ensures that `update_message` events won't come to us
with an empty string for `orig_subject`.

[1] https://chat.zulip.org/#narrow/stream/206-zulip-terminal/topic/subject.20always.20present.20in.20event/near/1098954
[2] https://github.com/zulip/zulip/blob/08d716c74175a8b63bf8eb1080381c77f3f853c6/zerver/views/message_edit.py#L157