podfile [nfc]: Set up Flipper Pods by calling a function exposed by RN.
Instead of spelling out the exact same code ourselves.

Corresponds to the portion of facebook/react-native@619d5d60d that
was missing in its backport to RN v0.62
(facebook/react-native@b4d1fcfb2); we took all of that incomplete
backport in 9a144c5b3.

After aligning our Podfile code with the corresponding code in the
RN v0.63 script, in #4322, this change can be done as an NFC commit
after the RN v0.63 upgrade commit.