api: Cut unused, unimplemented user-groups bindings.
Many of these can't possibly work in the first place -- how can we
edit a user group, for example, without providing any data on what
we want to change about it?

In any case, we don't use them, and have no plans to in the short or
medium term.  When we do, it's not hard to add bindings, and really
that will only be made easier by not having broken ones lying around.