user types: Use UserId for "own user ID" data.
This propagates the UserId type through where we store in Redux the
self user's user ID, to the `getOwnUserId` selector that gets that
information, and on to all the values fed by that selector.

We already have a pretty consistent convention of calling those values
`ownUserId`, so this is an easy change to make.