MainTabs [nfc]: Give screens a consistent naming convention.
It can get a bit hard to follow when, e.g.,

- one screen's name includes "Tab", like "HomeTab", to indicate it's
  the content you see when you've selected a particular tab

- another seems to include "Card", like "PmConversationsCard", to
  say the same thing

- another, "StreamTabs", uses neither; rather, "Tabs" there is meant
  to say that it's the screen with two tabs at the top, "All
  streams" and "Subscribed".

More of these say "Card" than not, so it's tempting to make that the
norm. However, I haven't found an indication that React Navigation
likes to use "card" in the context of bottom tab navigation [1].

The React Navigation docs are mixed on the question of whether to
use the "Screen" suffix; I can find examples with [2] and
without [3] it, and they specifically say, "The suffix `Screen` in
the component name is entirely optional, but a frequently used
convention" [4]. Might as well include it, for disambiguation among
potential similarly named things. And we've included it with most of
the things on `AppNavigator` (we'll include it on all of them in a
few commits).

Leave "Tabs" in "StreamTabsScreen" to indicate that it contains a
tab navigator.

[1] It *does* seem to be used for stack navigation. In my clone of
    the react-navigation repo at latest `main`, a case-insensitive
    search for "card" in packages/bottom-tabs shows two results,
    both from things defined elsewhere. A similar search in
    packages/stack shows 172 results, from things that mostly seem
    to be defined there.

[2] https://reactnavigation.org/docs/hello-react-navigation#creating-a-stack-navigator

[3] https://reactnavigation.org/docs/nesting-navigators

[4] https://reactnavigation.org/docs/glossary-of-terms/#screen-component