UserItem: Don't pass a do-nothing onPress down to the touchables.
When there's no `onPress` callback passed in to this component,
our own `handlePress` will do nothing.  Rather than pass that on
down as an `onPress` prop, pass nothing again.

This is useful because when our `Touchable` doesn't receive an
actual callback for interaction, it can avoid creating an
underlying `TouchableHighlight` or `TouchableNativeFeedback` at
all, and so avoid falsely advertising to a screen reader or other
a11y interface that there's some interaction being offered here.