internalLinks: Use purely user IDs in getNarrowFromLink for PM links.
See previous commit for detailed discussion.  This takes care of
the other of the two warnings we've been seeing recently when
running our tests.

Here again, the warning is logged by our own code when trying to
construct a PM narrow and we find that one of the mentioned users
doesn't exist in our data structures; and again the test that
evokes it is a test specifically for this error case.

And here again, the code in question doesn't actually need the
users at all and we can simplify by having it deal purely with user
IDs, making the code not even have to get passed the all-users data
structure and making this test no longer needed.

Here again, this means that if we do see the case that previously
would have caused a warning, we'll unsuspectingly return a Narrow
with the provided user ID in it, where previously we would have
returned null; and again that means we'll go ahead and navigate to
the specified narrow, and end up showing a reasonable error message
in ChatScreen.

In this case, the old behavior (based on my reading of this code's
caller, `messageLinkPress`, just now) would have ended up
attempting to open the broken link in a browser.  That seems
definitely worse: at best it'd end up with a comparable error
message in the browser, which is just unnecessarily roundabout;
and the most likely experience for most users is that they'd get a
login prompt, and either go through some hassle to find the right
password and log in, or give up and not know what information they
were missing behind the link.