safe-area: Remove an unnecessary condition on `orientation`.
In each of these instances, the `View` will be rendered at the top
of the screen -- regardless of the device's orientation. (I.e., if
it's in a portrait orientation, it'll be on a short side at the top;
if it's in a landscape orientation, it'll be on a long side at the
top.) The `View` is meant to guard against a potential unsafe area
at the top. In principle, that isn't less necessary when we're in a
landscape orientation.

I can't think of a device that has a nonzero top inset in landscape
mode, so it's possible that this commit is NFC. Still, we might as
well (1) handle that case, so we don't have to think about it, and
(2) remove something that can cause confusion.