ChatScreen [nfc]: Move safe-area-handling `View` into `ChatNavBar`.
The nav bar is always present, for `ChatScreen`, and always at the
top. So it's harmless to have the nav bar component also take
control of handling the upper safe-area inset. (The default
algorithm used to display the diff in `ChatNavBar` gives a lot of
noise; `git diff -w` should reduce that quite a lot.)

Soon, I'd like to start reversing our avoidance of React
Navigation's screen-header APIs; i.e., it'll make sense for the
`<Stack.Screen` with `component={ChatScreen}` to use `ChatNavBar`
(or something much like it) in its `header` prop [1], instead of
sticking it at the top of the screen component, where the screen's
content belongs.

When we do that, `ChatScreen` will properly describe only the
content of the screen, below the header, and it'll be too far down
the screen for top-safe-area handling to be done there.

[1] https://reactnavigation.org/docs/stack-navigator/#header