unread: Speed up initial-data loading by constructing in batch.
This substantially reduces the time we spend in constructing this
data structure at the initial fetch: from about 100ms, it goes down
to about 55ms.

The basic principle here is that if you have a whole bunch of items
you're going to build a data structure out of, it's faster to
construct the data structure with all of them at once than to start
with an empty data structure and add them one at a time.  That's
already what we routinely do in simple cases -- and in fact there's
a typical example at the innermost part of this code, the
`Immutable.List(unread_message_ids)`.

For this data structure as a whole (the Map of Maps of Lists), the
reason we didn't was just that the input comes to us as individual
topics, not grouped by stream.  But we can group them as a
preprocessing step, and then construct each Immutable data structure
all in one go without updates.  For the preprocessing step, we do need
a data structure we build up incrementally... but we can use plain
mutable built-in data structures for that, as it's temporary and
local; and even use just a plain Array for each stream's topics,
rather than any kind of Map, because we're only going to run through
the whole thing and never try to look up a particular item, or update
a particular existing item in it.  Those have a lot less overhead than
Immutable.Map, partly because they're built into the JS engine and
partly because an Array is a much simpler data structure than a map.