deps: Update react-native-vector-icons to 8.0.0, the latest.
As with the upgrade to v7, in 5ecf10ed1, only one breaking change
was announced in 8.0.0 [1], and it's for something we don't use:

- `Icon.TabBarItemIOS` convenience component removed as it was
  removed from React Native core.

It also advertises better compatibility with Xcode 12. I hadn't run
into any problems with Xcode 12, but still, this sounds helpful.

So, might as well take the latest version.

[1] https://github.com/oblador/react-native-vector-icons/releases/tag/v8.0.0