msglist: Gracefully ignore ill-formed image URLs.
Recently I've occasionally been seeing a red error banner at the top
of the message list (the one that says "Oh no! Something went
wrong. Try again?".)  I've only seen it when looking at #**test here**
on czo -- so figured there must be an oddity in some message there
that triggered whatever bug it was.

And indeed, it turns out that the trigger is a message with an image
element where the image's URL isn't a well-formed URL.  We'd been
implicitly assuming that that couldn't happen -- the server is after
all generating this HTML, and could prevent that.  But it turns out
that it doesn't.

Quite possibly the server should and in the future will refuse to
generate such elements.  But the Zulip server renders messages to HTML
immediately at send time, and in general we never go back and
re-render old messages.  (That's because if we did, then every subtle
change in Zulip's Markdown format, or even changes in its
implementation that were intended to have no effect on behavior, would
carry a risk of silently corrupting old messages.)  So even if the
server becomes stricter about this in the future, clients will still
need to handle messages with this kind of malformedness indefinitely.

(For that matter, in fact, it's possible the server already has made
that change in the two years since the message I found that triggers
this issue was sent.)

From the code, it looks like this bug has been here since we started
explicitly absolutizing these URLs, in 7ef0aebe0 in 2019.  Yet I was
pretty confident I hadn't seen it before, when I saw it sometime in
the last few weeks; and every time I've seen it since then, it's been
on the same stream and likely due to the same one message there.
So that's a little reassuring that this isn't a super common issue
to hit, at least.

Fixes: #4449