LoadingScreen [nfc]: Stop using as a navigation route, rename.
We recently stopped navigating to the 'loading' route at all, which
means we can stop listing it in `AppNavigator`. When we do so,
`LoadingScreen` loses all its former claims to a special status as a
React Navigation screen component. So, remove `route` and
`navigation` altogether from its props, rename the component, and
put it in src/common along with our other loading-indicator
components.