migrations: Quick fix for #4458.
See Greg's explanation for this quick fix at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4458.3A.20.22t.2Eget.20is.20not.20a.20function.22.20on.20state.2Enarrows.20at.20sta.2E.2E.2E/near/1119547
and his diagnosis and a further solution just above that.

We're pretty confident that this will fix the issue; we'll have to
monitor Sentry after we make a beta release to remove all doubt. If
not, we can always reopen the issue.

Fixes: #4458