ComposeBox: More correctly set the topic input when starting an edit.
As the existing code expresses, we want the topic input to populate
with the selected message's topic when we start editing a message in
a stream narrow.

However, "in a stream narrow" is too restrictive -- shouldn't we
also care about doing this in a topic narrow?

Well, yes, we should. The topic input does seem to get populated
when starting an edit in a topic narrow -- but in a *really*
convoluted way. In particular:

1. We set `state.topic` to the empty string, via
   `this.setTopicInputValue`.

2. Later on in `UNSAFE_componentWillReceiveProps`, we ask the
   message input to focus.

3. That, in turn, triggers `handleMessageFocus`.

4. That calls `setState` to update `topic` with `lastMessageTopic`
   (the `state.topic` branch having been skipped because we've just
   set that to be the empty string in step 1).

5. `lastMessageTopic` happens to be the right topic because all
   messages in a topic narrow have the same topic.

I'm fairly confident that this is what's happening. Having replaced
`|| lastMessageTopic` in `handleMessageFocus` from step 3 with
`|| 'hello world'`, I see the topic input populating with 'hello
world'.

How about opening it up for more than just stream and topic narrows
(as done in this commit)?

Well, whichever narrow it is, the selected message will either be a
stream message or a private message. We don't have the whole message
handy to check its `.type` for that, but that doesn't matter:

- For a private message, `editMessage.topic` will be the empty
  string (since all PMs have the empty string at `.subject`), so
  this change is NFC on PMs in narrows other than stream narrows
  (which is to say, it's NFC on all PMs).

- For a stream message in a narrow that's not a stream or topic
  narrow, `editMessage.topic` will be the message's topic, and we'll
  want to use that. In fact, this feels like a bugfix for such a
  message -- in the previous code, number 5, above, wouldn't hold,
  and the wrong topic might have been chosen!