modelTypes: Split `Message.type` into `PmMessage` and `StreamMessage`.
This is the first thing that makes `PmMessage` and `StreamMessage`
differ.

The fact that we don't see Flow errors at this commit is a
reasonable sign that we're ready to proceed with giving this
treatment to the other pm- and stream-message-specific fields. Flow
will likely give us substantive feedback when we do that, so we'll
take it one field at a time.