flow: Remove `suppress_comment` for `$FlowMigrationFudge`.
And change its uses to `$FlowFixMe`.

`suppress_comment` is removed in Flow v0.127.0 [1], and we'll only
be able to use the built-in suppression syntax from that version on.

`$FlowFixMe` isn't the ideal replacement for `$FlowMigrationFudge`;
`$FlowIgnore` would match the meaning more closely. But
`$FlowIgnore` doesn't get added to the built-in suppression syntax
until Flow v0.128.0 (facebook/flow@ba78ab9bf).

Fortunately, it seems quite straightforward to take Flow v0.128
before the main RN v0.64 upgrade commit. So, we'll do that later in
this series, and switch these uses to `$FlowIgnore` afterward.

[1] https://flow.org/en/docs/config/options/#toc-suppress-comment-regex