flow: Remove all `suppress_comment`s.
Now, we're not erasing any of the default suppression comments, and
we're not adding our own. So at this point we're relying on exactly
the default set of suppression comments. This is a requirement as of
Flow v0.127. [1]

Follows the template-app change in facebook/react-native@afad239d7,
on the path to the RN v0.64 upgrade.

[1] https://flow.org/en/docs/config/options/#toc-suppress-comment-regex