webview: Update Chrome version threshold for graceful degradation.
When we first introduced this threshold in 1298a4f7b, it was 2018-10
and Android 8 was just over a year old.  It's now approaching 2.5
years later, and I think we can safely advance this threshold by a
couple of years.

Realistically we're going to want to advance it for sure: in fact
we've already had things broken for a while on older than Chrome 62,
thanks to #4451.  When we fix that so it doesn't cause any collateral
damage, we're still going to want to leave that particular Zulip
feature (namely "mention a time") degraded on those older versions.

Also add some text explaining this policy in general, and an update
of our notes on which Chrome versions come in stock emulator images
for which versions of Android.