deps: Downgrade react-native-vector-icons to 7.1.0.
From 8.0.0, the release that followed 7.1.0.

On the way down to 7.0.0, to avoid
oblador/react-native-vector-icons@269bce013 (released in 7.1.0),
which caused certain Feather icons to look wrong; that's our #4521.

This cleanly reverts 4bc42a9e3. As noted there, there weren't
reasons we chose to go to ^8.0.0, beyond 8.0.0 being the latest
version at the time.

The changelog [1] has this for 8.0.0:

"""
Breaking changes:

- `Icon.TabBarItemIOS` convenience component removed as it was
  removed from React Native core.

Improvements:

- Improved Xcode 12 compatibility
"""

That bit about Xcode 12 compatibility might end up being a concern,
but I haven't so far observed any difficulty building or running
with Xcode 12.4 at 7.1.0, in debug or release modes. The changelog
doesn't go into detail, but I believe
oblador/react-native-vector-icons#1251 is meant.

[1] https://github.com/oblador/react-native-vector-icons/releases/tag/v8.0.0