events: Catch exceptions separately for handling events vs polling.
These are pretty different kinds of errors.

An exception while making a request may be a server bug, operational
issue, or just a lack of network connection; an exception while
handling an event is definitely our bug (or I guess potentially in
some library we use.)

And the appropriate responses are quite different: there's no sense
in setting a backoff timer on account of an exception in our handling
of an event.