unread [nfc]: Add much more jsdoc on the unreads data.
The jsdoc on the overall `UnreadState` hadn't been updated to reflect
that `UnreadStreamsState` is now our own data structure rather than
the straight JSON-decoding of the wire format.  We also didn't have
written down anywhere clearly what the keys and values are in
`UnreadStreamsState` -- it's not too hard to guess, but one shouldn't
have to guess.