FullScreenLoading [nfc]: Move `ZulipStatusBar` out of a `View`.
This departure from a normal parent-child `View` layout hierarchy in
a component is an imperfect way to emphasize that `ZulipStatusBar`
isn't meant to participate in that hierarchy at all; its
presence/absence with particular props is just a signal to the
native status bar API about how the status bar should behave.

Possibly we should make a HOC `withStatusBarSettings` that
encapsulates details like this.