contentHtmlFromPieceDescriptors tests: Add, with snapshot tests.
It seems worth it to try snapshot testing. We've run into the idea
before [1], and this time it actually seems like it might be useful.

See Jest's documentation on snapshot testing:
  https://jestjs.io/docs/snapshot-testing

Possibly, the new tests will fail too often, cause too much noise,
and slow down development. Possibly, they'll be great for catching
subtle bugs before they get merged. The next commit in this series
demonstrates a trivial change to the logic and how the snapshots
must be updated to follow.

The immediate motivation is to help with an upcoming refactor of
`getHtmlPieceDescriptors` and `contentHtmlFromPieceDescriptors`. We
hand data (messages, flags, etc.) to the former, which hands its
output to the latter, which -- we hope -- gives the right content
HTML string. It's an important area to get right, and the code is
tricky, so we need this coverage in place before doing the refactor.
If we're skeptical about the tests' long-term value, we can
cheerfully remove them, but maybe wait on that until they've done
their job for this refactor.

[1] https://github.com/zulip/zulip-mobile/pull/4348#discussion_r548762941