unread: Revert "Use state.messages to optimize updating stream unreads."
This reverts commit 8dc3fe605c5ee305555c515b2f938c83bd6d4ce8.

This optimization has a bug: to locate a message in the unreads state,
it looks it up in the messages state, but the messages state may be
far less complete than the unreads state.  (Which is in fact the point
of the server feature that gives us the initial data to populate the
unreads state.)  So the message may be absent in `state.messages`, in
which case we'll fail to remove it from `state.unread`.

The symptom is that if you have the app running, then read some
messages from the webapp or another client, they won't disappear from
the unreads view.

For now, just revert this optimization.  We're still in a better shape
than we were with the old data structures.

Fundamentally we have all the data we need for this optimization -- if
the message ID is in the unreads state, then we have the data of where
in the unreads state it belongs -- and just need a data structure
where that's indexed by message ID.  I've filed #4684 for doing that.
But fix the bug first.

See discussion:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/syncing.20read.2Funread.20state/near/1166265

While here, also fix a small bug in the `updateAllAndPrune` helper
from before the reverted commit: add a return after the `delete` call.
The existing tests catch this; I'd refactored that function in a
revision shortly before merge, and apparently didn't retest the
intermediate commits in the branch.

Reported-by: Wesley Aptekar-Cassels <me@wesleyac.com>