migrations: Remove accounts with empty `.email`.
I think the effect of not having this migration, at this point,
isn't nearly as bad as a crash. But we should have added such a
migration in 4efdcda7a (#4491), when we removed the only mechanism
(the REALM_ADD action) by which accounts could end up with the empty
string at `.email`. There, we implicitly added an invariant that no
accounts would ever have an empty `.email`. If I'd made this more
explicit, I might have remembered to do the migration.

Luckily, 1af58afb1 (from later in the same PR) ensured that we'd
never take an account object with an empty-string `.email` and make
it the active account. After #4491, if you tapped an
in-progress-account item in the list, typed good credentials, and
logged in successfully, a new account object would get created, with
non-empty `.email` (and `.apiKey`). The one with the empty `.email`
would sit harmlessly in the accounts list, and the user could delete
it if they want.

I verified the non-crash at the tip of #4491, b570b262c.