tools/test: Add `--platform` option.
And expand `--all`, which runs in CI and at releases, to include the
behavior of `--platform both`.

Leave the current behavior as the default, with the name "sloppy",
to not slow down local testing.

Greg points out [1],

"""
From looking at a couple of recent CI builds, it looks like running
Jest both ways doubles the Jest time, as you'd expect; it goes from
about 60s to about 120s.

Our CI times are around 8-9 minutes now, so that adds one minute --
not awesome, but I think worth it for getting coverage of this sort
of code.
"""

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60jest-expo.2Fios.60.20and.20.60jest-expo.2Fandroid.60/near/1169341