muted users tests: Use eg.otherUser and eg.thirdUser.
This helps make it explicit that there's nothing special about these
particular user IDs, other than the fact that one appears in certain
places in these tests and the other appears in certain other places.

It also means that if in the future we were to have any code that
expects these user IDs to be the IDs of users that we actually know
about -- which is an invariant that it'd be quite reasonable for code
to rely on -- these tests will be readily compatible with that because
there are already full User objects for these.