ZulipAsyncStorage: Also call callback on `AsyncStorage.getItem` failure.
A bug like this is another reason I'm continuing to think we should
just not bother to maintain a callback-based interface, here and in
`.setItem`, and just maintain the Promise-based interface.

Before we vendored redux-persist, which is the home of these
methods' only callers, this would have been trickier. But now it's
easy to make those callers adapt to the Promise-based interface.