api [nfc]: Fix remaining few broken links to API docs.
Found with this command, similar to the previous commit:

  $ git grep -hoE "https://(zulip(chat)?.com|chat.zulip.org)/api/[a-z-]+" \
    | xargs -n1 -P8 \
        curl -s -o /dev/null -w "%{http_code}\t%{url_effective}\n" \
    | grep -v ^200

The outright broken links were due to all the names in the docs
being changed a while ago in an effort to make them more
regularized.  Apparently when we did that we didn't leave
redirects from the old URLs. :-/

There were also a handful pointing to zulipchat.com, which redirects
to zulip.com; so while we're at it, follow those redirects.