ios: Remove dead code for supporting Apple's "Universal Links".
Apple's doc describes the feature this way [1]:

> When you support universal links, iOS users can tap a link to your
> website and get seamlessly redirected to your installed app
> without going through Safari. If your app isn’t installed, tapping
> a link to your website opens your website in Safari.
>
> Universal links give you several key benefits that you don’t get
> when you use custom URL schemes. [...]

React Native's `Linking` doc [1] says this code is needed if your
app is using Universal Links. (They probably also mean that you
don't need it if your app isn't using Universal Links.) Our app
doesn't use that, and it probably won't for quite some time. So, cut
out this dead code, with a plan to do whatever is the appropriate
setup for supporting the feature at the time we choose to do so.

[1] https://developer.apple.com/library/archive/documentation/General/Conceptual/AppSearch/UniversalLinks.html
[2] https://reactnative.dev/docs/0.63/linking