fetchActions: Fail early on all non–known-retryable errors in `tryFetch`.
The TODO doesn't specifically mention that we want to retry on
network-failure errors, but we do, so add that too.

If we get some strange, totally unexpected error caused by a bug
somewhere, we shouldn't continue the retry loop. Not only would that
waste the user's time, but it means the error basically gets
swallowed, and it makes it much harder for us to detect it and debug
effectively. So, fail early on exceptions like that.

See our doc on exception handling:
  https://github.com/zulip/zulip-mobile/blob/master/docs/style.md#catch-at-ui.