android notif: Make `getInitialNotification` a consumer.
When an app quits android system does not flush the entire memory
of the app until absolutely necessary [1] (possibly to facilitate
faster startup next time). This also involves the static variable
`initialNotification`.

This situation creates a problem when the app is started next time,
where it processes `initialNotification` again and falsely navigates
to its narrow.

[1] https://developer.android.com/guide/components/activities/activity-lifecycle#alc

Fixes: #4758