jest config [nfc]: Clarify organizing-comments a bit.
These comment lines are each meant to label a group of several
properties in the config: see 0f829fe78 where they were introduced.

Originally that was done by having each heading as a single line
attached to a stanza.  That perhaps worked when first introduced,
but it promptly stopped working at cc5d38211, when one stanza was
broken up to make room for a longer comment on a particular
property.  The comment no longer looked like a heading for a group
of lines, and as a result in b9d3dbae5 it got attached to just the
one line right after it.

So, try making the headings bigger, in particular to try to make it
clear that their scope is at a higher level than the comments on
individual properties in each section.